fix tracebacks
