+++ pytest_robotframework/_internal/patches/__init__.py
@@ -1,0 +1,4 @@
+"""these are the the hackiest parts of pytest-robotframework. they involve replacing some of pytest
+and robotframework's internals to do what we need.
+
+these modules mutate other modules as they're imported"""
