+++ tests/fixtures/test_python/test_full_stack_keyword_decorator.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from pytest_robotframework import keyword
+
+
+@keyword
+def bar():
+    raise Exception
+
+
+def test_keyword():
+    bar()
